+++ src/pages/create.tsx
@@ -1,39 +1,13 @@
 import { NextPage } from 'next'
 import Head from 'next/head'
-import { useRouter } from 'next/router'
-import { useCallback } from 'react'
 import { PostForm } from 'components/post/PostForm'
 import { Layout } from 'components/shared/Layout'
-import { useAuthHeaderParams } from 'hooks/login/useAuth'
 import { useRequireLogin } from 'hooks/login/useRequireLogin'
-import { useGetApi } from 'hooks/useApi'
-import { Post, PostRequestParams } from 'types/post'
-import { HttpError, postApi } from 'utils/api'
+import { useCreatePost } from 'hooks/post/useCreatePost'
 
 const Create: NextPage = () => {
   useRequireLogin()
-
-  const router = useRouter()
-  const authHeaderParams = useAuthHeaderParams()
-  const { data: posts, mutate } = useGetApi<Post[]>('/posts')
-
-  const onSubmit = useCallback(
-    async (params: PostRequestParams) => {
-      try {
-        const newPost = await postApi<Post>('/posts', params, authHeaderParams)
-        if (newPost && posts) {
-          console.log('投稿の作成に成功 ', newPost)
-          router.push('/')
-          mutate([...posts, newPost], false)
-        }
-      } catch (e) {
-        if (e instanceof HttpError) {
-          console.error(e.message)
-        }
-      }
-    },
-    [authHeaderParams, mutate, posts, router],
-  )
+  const { createPost } = useCreatePost()
 
   return (
     <>
@@ -43,7 +17,7 @@ const Create: NextPage = () => {
       <Layout>
         <h1 className='text-center text-lg'>記事投稿</h1>
         <div className='mt-4'>
-          <PostForm onSubmit={onSubmit} />
+          <PostForm onSubmit={createPost} />
         </div>
       </Layout>
     </>
